Westeros/requirements.txt
django-ninja==0.21.0
django==4.1.7
